requirements.txt
PyJWT>=2.8.0
cryptography>=41.0.0
djangorestframework>=3.14.0
django>=4.2.0
lxml>=4.9.0
signxml>=3.1.0
requests>=2.31.0
